make_money

+++ controllers/expense.js
@@ -1,5 +1,5 @@
 const Expense = require('../models/expenses');
-const addexpense = (req,res) => {
+const addExpense = (req,res) => {
     const {amount, desc, category} = req.body;
     if(amount == undefined || amount.length == 0){
         return res.status(400).json({success: false, message: 'parameter is missing'})
@@ -18,21 +18,23 @@ const getexpenses = (req, res) => {
 return res.status(402).json({error:err, success: false})
     })
 }
-const deleteexpense = (req,res) => {
+const deleteExpense = (req,res) => {
    const expenseid = req.params.expenseid;
    if(expenseid == undefined || expenseid.length == 0){
        res.status(400).json({success: false})
    }
-   Expense.destroy({where: {id: expenseid}}).then(() => {
-     return res.status(204).json({success: true, message:" Deleted Successfully"})
+   Expense.destroy({where: {id: expenseid, userId: req.user.id}}).then((noofrows) => {
+    if(noofrows == 0){
+        return res.status(404).json({success: false, message:'Expense does not belong to the user'})
+    }
+     return res.status(200).json({success: true, message:" Deleted Successfully"})
    }).catch(err => {
     console.log(err);
     return res.status(403).json({success: true, message:"failed"})
    })
 }
 module.exports = {
-    deleteexpense,
+    addExpense,
     getexpenses,
-    addexpense
-
+    deleteExpense
 }
